Index * create superclass for index * point subclass to superclass * remove query method * remove TODO
