+++ index/conjunctive_index.py
@@ -1,37 +1,14 @@
 """Module for a conjunctive index"""
-import os
-import json
 from queue import PriorityQueue
 
+from .index import Index
 
-# TODO create Index super class
-class ConjunctiveIndex:
+
+class ConjunctiveIndex(Index):
     """Class for building a conjunctive index."""
 
     def __init__(self):
         self.index = {}
-
-    # TODO make the path check into a decorator
-    def load(self, path):
-        """Loads a stored conjunctive index."""
-        if not os.path.isdir(path):
-            raise OSError(f'The path: {path} does not exist.')
-
-        file_name = 'conjunctive_index.json'
-        full_path = f'{path}/{file_name}'
-
-        with open(full_path, 'r') as file:
-            self.index = json.load(file)
-
-    def save(self, path):
-        """Saves the conjunctive index into a file."""
-        if not os.path.isdir(path):
-            raise OSError(f'The path: {path} does not exist.')
-
-        file_name = 'conjunctive_index.json'
-        full_path = f'{path}/{file_name}'
-        with open(full_path, 'w') as file:
-            json.dump(self.index, file)
 
     def put(self, word_id, document: str):
         """Puts a document into the inverted index at the word_id.
@@ -62,6 +39,3 @@ class ConjunctiveIndex:
 
         except KeyError:
             raise KeyError(f'The word: {word_id}, does not exist in the index.')
-
-    def query(self, query_words):
-        pass
